pyproject.toml
[project]
name = "hl-personal"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "hyperliquid-python-sdk>=0.22.0",
    "python-dotenv>=1.2.1",
]
